+++ docs/source/manual/tutorials/ollama_benchmark/pipelines.py
@@ -124,6 +124,6 @@ def ollama_benchmark_pipeline():
     nodes = {
         'ollama_benchmark': OllamaBenchmark(),
     }
-    dag = kwdagger.Pipeline(nodes)
+    dag = kwdagger.Pipeline(list(nodes.values()))
     dag.build_nx_graphs()
     return dag