+++ kwdagger/aggregate.py
@@ -2210,7 +2210,7 @@ class Aggregator(
         agg._metric_info = {}
 
         if agg.dag is not None:
-            node = agg.dag.nodes[agg.node_type]
+            node = agg.dag.node_dict[agg.node_type]
         else:
             node = None
 
@@ -2435,7 +2435,7 @@ class Aggregator(
         try:
             assert self.node_type is not None
             if self.dag is not None:
-                node = self.dag.nodes[self.node_type]
+                node = self.dag.node_dict[self.node_type]
                 vantage_points = node.default_vantage_points
         except Exception:
             vantage_points = []